ONB-88 fix: Group File Url -> File Entity
