+++ domain/src/main/kotlin/com/yapp/bol/group/Group.kt
@@ -3,6 +3,7 @@ package com.yapp.bol.group
 import com.yapp.bol.InvalidGroupDescriptionException
 import com.yapp.bol.InvalidGroupNameException
 import com.yapp.bol.InvalidGroupOrganizationException
+import com.yapp.bol.file.FileInfo
 
 @JvmInline
 value class GroupId(val value: Long)
@@ -12,7 +13,7 @@ data class Group(
     override val name: String,
     override val description: String,
     override val organization: String?,
-    override val profileImageUrl: String = DEFAULT_PROFILE_IMAGE_URL,
+    override val profileImage: FileInfo,
     override val accessCode: String = generateAccessCode(),
 ) : GroupBasicInfo {
     init {
@@ -55,6 +56,6 @@ interface GroupBasicInfo {
     val name: String
     val description: String
     val organization: String?
-    val profileImageUrl: String
+    val profileImage: FileInfo
     val accessCode: String
 }
